Implementing a resource registry in python side to render includes automatically
Now, instead of adding includes directly to a template, we should manage them
in the `resources.json' file. This file is separated by types of resources (currently,
css and js) and by groups.

These groups can be used to say which files should be merged to serve in a given
template. But as the `gallery' group, you can groups things by its features and
include more than one group in the template.